+++ app/models/OrderLike.scala
@@ -2,8 +2,6 @@ package models
 
 import akka.actor.ActorRef
 
-import scala.concurrent.duration.FiniteDuration
-
 
 /** Trait representing an order for a particular security.
   *
@@ -20,10 +18,10 @@ trait OrderLike {
   def instrument: AssetLike
 
   /** The quantity being bought or sold. */
-  def quantity: Int
+  def quantity: Double
 
   /** Orders will often need to be split during the matching process. */
-  def split(newQuantity: Int): OrderLike
+  def split(newQuantity: Double): OrderLike
 
   /** String representation of an order. */
   def toString: String
